+++ CHANGELOG.md
@@ -9,16 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0).
 
 ### Added
 
-- `initialize` to set up the data directory (including a readme and the search_details)
+- `initialize` to set up the data directory (including a `readme.md` and the `search_details.csv`)
 
 ### Changed
 
 - Refactored code, including `analysis/utils.py/load_references_bib(modification_check, initialize)` and `analysis/utils.py/git_modification_check(filename)`
+- Improve treatment of diacritics and accents when generating citation_keys in `analysis/cleanse_records.py`
 
 ### Fixed
 
 - Bugs in `analysis/combine_individual_search_results.py` and in `analysis/acquire_pdfs.py`
 - Catch exceptions and check bad responses in `analysis/acquire_pdfs.py`
+- Bug in git modification check for `references.bib` in `analysis/utils.py`
 
 ### [0.1.0] -2021-05-08